pyproject.toml
[project]
name = "financial-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.3",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-google-community>=3.0.4",
    "langchain-google-genai>=4.1.3",
    "langchain-google-vertexai>=3.2.1",
    "python-dotenv>=1.2.1",
]
